Added proper reader stream support
This allows the reader to properly handle reading data from a blocking socket. When the reader needs data it will request filling the buffer, but the fill function can always return less than the requested amount as long as it read at least one byte. This prevents the reader from blocking at the end of a message, or of causing errors on messages split between multiple packets.