Moved tree timestamp check to usage
The format of supported extension types (e.g. timestamp) is now checked lazily. An error is only raised on invalid timestamps if a timestamp function is actually called.

This fixes a backwards compatibility issue where incompatible ext types would be flagged as errors before allowing the application to interpret them. From now on ext types will never flag an error on their content unless MPack is asked to interpret it.