Add raw float/double support, double-to-float conversion
Floats and doubles can now be parsed as raw data if the platform does
not support `float` or `double`. On platforms that support only `float`
and not `double`, this implements a manual conversion from a 64-bit
double to `float` to allow parsing doubles as floats. (See #79 for a bit
of discussion on this.)